+++ stevelib/src/main/java/com/pheiffware/andpheifflib/physics/entity/rigidBody/PolygonEntity.java
@@ -5,11 +5,11 @@
 package com.pheiffware.andpheifflib.physics.entity.rigidBody;
 
 
-import com.pheiffware.andpheifflib.geometry.Vec3D;
-import com.pheiffware.andpheifflib.geometry.intersect.IntersectCalc;
-import com.pheiffware.andpheifflib.geometry.intersect.IntersectionInfo;
-import com.pheiffware.andpheifflib.geometry.shapes.OrientedLineSegment;
-import com.pheiffware.andpheifflib.geometry.shapes.Sphere;
+import com.pheiffware.andpheifflib.geometry.d3.Vec3D;
+import com.pheiffware.andpheifflib.geometry.d3.intersect.IntersectCalc;
+import com.pheiffware.andpheifflib.geometry.d3.intersect.IntersectionInfo;
+import com.pheiffware.andpheifflib.geometry.d3.shapes.OrientedLineSegment;
+import com.pheiffware.andpheifflib.geometry.d3.shapes.Sphere;
 import com.pheiffware.andpheifflib.physics.InteractionException;
 import com.pheiffware.andpheifflib.physics.entity.physicalEntity.PhysicalEntity;
 import com.pheiffware.andpheifflib.physics.entity.physicalEntity.PhysicalEntityCollision;
@@ -101,7 +101,7 @@ public class PolygonEntity extends PhysicalEntity
 		}
 		boundingSphere = new BoundingSphere(
 				Vec3D.scale(Vec3D.add(p1, p2), 0.5f),
-				(double) Math.sqrt(longestDistanceSquared) / 2);
+				Math.sqrt(longestDistanceSquared) / 2);
 	}
 
 	/*
